Refactored the code into modules, setting up cron schedule so it can be automated, will be adding docker in next branch
