+++ .env.example
@@ -1,2 +1,6 @@
 apiKey='Put your hevy API key here'
 baseUrl='https://api.hevyapp.com/v1/'
+# This schedule will run the job every day at 1 AM, delete schedule if you want to only run on demand
+schedule=0 1 * * *
+#Delete this variable if you want it to use kg
+useLbs=true
