+++ package.json
@@ -11,7 +11,8 @@
   "type": "module",
   "dependencies": {
     "cross-env": "^7.0.3",
-    "dotenv": "^16.5.0"
+    "dotenv": "^16.5.0",
+    "node-cron": "^3.0.3"
   },
   "repository": {
     "type": "git",
